+++ permission_handler_html/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+
+* Fixed plugin initialization for non-https web app.
+* Fixed location permission name
+* Improved error handling in the example app
+
 ## 0.1.0+1
 
 * Updates `permission_handler_platform_interface` dependency to version `^4.0.2`.